Rename package docker -> container
