+++ node-admin/src/main/java/com/yahoo/vespa/hosted/node/admin/container/package-info.java
@@ -1,5 +1,5 @@
 // Copyright 2018 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 @ExportPackage
-package com.yahoo.vespa.hosted.node.admin.docker;
+package com.yahoo.vespa.hosted.node.admin.container;
 
 import com.yahoo.osgi.annotation.ExportPackage;
